Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -1,7 +1,7 @@
 #include<stdio.h>
 /**
 * main - A programm that prints the size of various types on the computer
-* Return 0 (Success)
+* Return: Always 0 (Success)
 */
 int main(void)
 {
@@ -10,10 +10,10 @@ int b;
 long int c;
 long long int d;
 float f;
-printf("Size of char: %ld byte\n", sizeof(a));
-printf("Size of int: %ld bytes\n", sizeof(b));
-printf("Size of long int: %ld bytes\n", sizeof(c));
-printf("Size of long long int: %ld bytes\n", sizeof(d));
-printf("Size of float: %ld byte\n", sizeof(f));
+printf("Size of char: %ld byte\n", (unsigned long)sizeof(a));
+printf("Size of int: %ld bytes\n", (unsigned long)sizeof(b));
+printf("Size of long int: %ld bytes\n", (unsigned long)sizeof(c));
+printf("Size of long long int: %ld bytes\n", (unsigned long)sizeof(d));
+printf("Size of float: %ld byte\n", (unsigned long)sizeof(f));
 return (0);
 }
